Move nav into it's own component.
