+++ src/Components/Nav/Nav.css
@@ -1,0 +1,28 @@
+header {
+  background-color: #D9D9D9;
+  padding: 1em 2em;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+header > h1 {
+  margin: 0;
+  font-family: 'Antic Didone', sans-serif;
+  font-size: 2.5rem;
+  font-weight: 200;
+  color: #000;
+}
+
+header > nav {
+  display: flex;
+  gap: 2rem;
+}
+
+.nav-link {
+  font-family: 'Josefin Sans', sans-serif;
+  text-decoration: none;
+  font-weight: 200;
+  font-size: 1.5rem;
+  color: #000;
+}
